# API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Tranga v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/Chapters/Manga/{MangaId}": {
      "post": {
        "tags": [
          "Chapters"
        ],
        "summary": "Returns all API.Schema.MangaContext.Chapter of API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page to request (default 1)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Size of Page (default 10)",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/ChapterFilterRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterPagedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Page data wrong",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database request"
          }
        }
      }
    },
...
    "/v2/Manga/{MangaId}": {
      "get": {
        "tags": [
          "Manga"
        ],
        "summary": "Return API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Manga"
        ],
        "summary": "Delete API.Controllers.DTOs.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Controllers.DTOs.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "API.Controllers.DTOs.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/MangaConnector/{MangaConnectorName}": {
      "get": {
        "tags": [
          "MangaConnector"
        ],
        "summary": "Returns the API.MangaConnectors.MangaConnector (Scanlation-Sites) with the requested Name",
        "parameters": [
          {
            "name": "MangaConnectorName",
            "in": "path",
            "description": "API.MangaConnectors.MangaConnector.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/MangaConnector"
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.MangaConnector (Scanlation-Sites) with Name not found.",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/MangaConnector/Enabled/{Enabled}": {
      "get": {
        "tags": [
          "MangaConnector"
        ],
        "summary": "Get all API.MangaConnectors.MangaConnector (Scanlation-Sites) with Enabled-Status",
        "parameters": [
          {
            "name": "Enabled",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MangaConnector"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/MangaConnector/{MangaConnectorName}/SetEnabled/{Enabled}": {
      "patch": {
        "tags": [
          "MangaConnector"
        ],
        "summary": "Enabled or disables API.MangaConnectors.MangaConnector (Scanlation-Sites) with Name",
        "parameters": [
          {
            "name": "MangaConnectorName",
            "in": "path",
            "description": "API.MangaConnectors.MangaConnector.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Enabled",
            "in": "path",
            "description": "Set true to enable, false to disable",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "API.MangaConnectors.MangaConnector (Scanlation-Sites) with Name not found.",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AltTitle": {
        "required": [
          "language",
          "title"
        ],
        "type": "object",
        "properties": {
          "language": {
            "minLength": 1,
            "type": "string",
            "description": "Language of the Title"
          },
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "Title"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.AltTitle DTO"
      },
      "Author": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Author."
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "The API.Schema.MangaContext.Author DTO"
      },
      "Chapter": {
        "required": [
          "chapterNumber",
          "downloaded",
          "key",
          "mangaConnectorIds",
          "mangaId",
          "title",
          "volume"
        ],
        "type": "object",
        "properties": {
          "mangaId": {
            "minLength": 1,
            "type": "string",
            "description": "Identifier of the Manga this Chapter belongs to"
          },
          "volume": {
            "type": "integer",
            "description": "Volume number",
            "format": "int32"
          },
          "chapterNumber": {
            "minLength": 1,
            "type": "string",
            "description": "Chapter number"
          },
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "Title of the Chapter"
          },
          "downloaded": {
            "type": "boolean",
            "description": "Whether Chapter is Downloaded (on disk)"
          },
          "mangaConnectorIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMangaConnectorId"
            },
            "description": "Ids of the Manga on MangaConnectors"
          },
          "fileName": {
            "type": "string",
            "description": "Filename of the archive",
            "nullable": true
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.Chapter DTO"
      },
      "ChapterFilterRecord": {
        "type": "object",
        "properties": {
          "downloaded": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "volumeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "chapterNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChapterMangaConnectorId": {
        "required": [
          "foreignKey",
          "key",
          "mangaConnectorName",
          "useForDownload"
        ],
        "type": "object",
        "properties": {
          "mangaConnectorName": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Connector"
          },
          "foreignKey": {
            "minLength": 1,
            "type": "string",
            "description": "Key of the referenced DTO"
          },
          "websiteUrl": {
            "type": "string",
            "description": "Website Link for reference, if any",
            "nullable": true
          },
          "useForDownload": {
            "type": "boolean",
            "description": "Whether this Link is used for downloads"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Controllers.DTOs.MangaConnectorId`1 DTO"
      },
      "ChapterPagedResponse": {
        "required": [
          "data",
          "page",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Chapter"
            },
            "description": "The data"
          },
          "page": {
            "type": "integer",
            "description": "The page-number",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages (max page)",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "The total number of entries for this request",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ContentKind": {
        "enum": [
          "Manga",
          "Novel"
        ],
        "type": "string"
      },
...
      "Link": {
        "required": [
          "key",
          "provider",
          "url"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Provider"
          },
          "url": {
            "minLength": 1,
            "type": "string",
            "description": "Url"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.Link DTO"
      },
      "Manga": {
        "required": [
          "altTitles",
          "authors",
          "description",
          "fileLibraryId",
          "ignoreChaptersBefore",
          "key",
          "links",
          "mangaConnectorIds",
          "name",
          "releaseStatus",
          "tags"
        ],
        "type": "object",
        "properties": {
          "ignoreChaptersBefore": {
            "type": "number",
            "description": "Chapter cutoff for Downloads (Chapters before this will not be downloaded)",
            "format": "float"
          },
          "year": {
            "type": "integer",
            "description": "Release Year",
            "format": "int32",
            "nullable": true
          },
          "originalLanguage": {
            "type": "string",
            "description": "Release Language",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "description": "Author-names"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Manga Tags"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "description": "Links for more Metadata"
          },
          "altTitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AltTitle"
            },
            "description": "Alt Titles of Manga"
          },
          "fileLibraryId": {
            "minLength": 1,
            "type": "string",
            "description": "Id of the Library the Manga gets downloaded to"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Manga"
          },
          "description": {
            "minLength": 1,
            "type": "string",
            "description": "Description of the Manga"
          },
          "releaseStatus": {
            "$ref": "#/components/schemas/MangaReleaseStatus"
          },
          "mangaConnectorIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaMangaConnectorId"
            },
            "description": "Ids of the Manga on MangaConnectors"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.Manga DTO"
      },
      "MangaConnector": {
        "required": [
          "contentKind",
          "enabled",
          "iconUrl",
          "key",
          "supportedLanguages"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether Connector is used for Searches and Downloads"
          },
          "supportedLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Languages supported by the Connector"
          },
          "iconUrl": {
            "minLength": 1,
            "type": "string",
            "description": "Url of the Website Icon"
          },
          "contentKind": {
            "$ref": "#/components/schemas/ContentKind"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false
      },
      "MangaMangaConnectorId": {
        "required": [
          "foreignKey",
          "key",
          "mangaConnectorName",
          "useForDownload"
        ],
        "type": "object",
        "properties": {
          "mangaConnectorName": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Connector"
          },
          "foreignKey": {
            "minLength": 1,
            "type": "string",
            "description": "Key of the referenced DTO"
          },
          "websiteUrl": {
            "type": "string",
            "description": "Website Link for reference, if any",
            "nullable": true
          },
          "useForDownload": {
            "type": "boolean",
            "description": "Whether this Link is used for downloads"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Controllers.DTOs.MangaConnectorId`1 DTO"
      },
      "MangaReleaseStatus": {
        "enum": [
          "Continuing",
          "Completed",
          "OnHiatus",
          "Cancelled",
          "Unreleased"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}